move layout code to app to reduce calls, remove cross tab sync which caused spamming of requests, fix address country, add warning to payment method fetching
