+++ components/Layout/Footer/LayoutFooterPaymentMethods.vue
@@ -5,6 +5,7 @@ import type { Schemas } from '@shopware/api-client/api-types';
 // We can't use the useCheckout composable to get the paymentMethods because it always uses the parameter 'onlyAvailable'
 // This means that it applies rules from e.g. the rule builder depending on the current context
 // We want to show ALL possible payment methods here though, not just the ones currently available for the customer / cart
+// This causes load on the server due to the api call being made - if the payment methods do not change frequently it is advised to define them statically
 const { apiClient } = useShopwareContext();
 const { data: paymentMethods } = await apiClient.invoke(
     'readPaymentMethod post /payment-method',
